refactor(stream-create): hoist handoff types to lib/
DesktopInitializer was importing from a feature module
(components/features/streams/create/create-stream-dialog) just to grab the
sessionStorage key + payload type. The dependency arrow was backwards — a
low-level initializer shouldn't know about a specific dialog component.

Move streamCreateHandoffKey + StreamCreateHandoff to lib/ so both producer
(the dialog) and consumer (DesktopInitializer) import from a neutral place.
